chartsense_backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0

# PDF and Image Processing (Pure Python, compatible with Python 3.12)
pymupdf>=1.23.0
pillow>=10.0.0
pdf2image>=1.17.0

# Hugging Face Integration
huggingface_hub>=0.19.0
requests>=2.31.0

# Additional utilities for table processing
numpy>=1.24.0
opencv-python-headless>=4.8.0